Fix problem saving to database
The problem happend because the solution name was null. I have made
sure that this can't happen, and an empty string is now used instead.

I have also found the reason why the solution name was null and
fixed that. The Opened event doesn't fire when a solution is
opened when starting vs 2019, as I think we only connect to the
events after the solution is already open. To get round this I call
the event manually during initialize.